Tidied Phasmo actions architecture.  Needs sorting for GUI UX
